mirror regions
