+++ src/atomic-orbital-viewer/data/orbital-data.js
@@ -148,7 +148,7 @@ let orbitals = [
         ]
     },
     {
-        minimum: 0,
+        minimum: -4.5,
         maximum: 4.5,
         regions: [
             {
@@ -162,6 +162,14 @@ let orbitals = [
                 isConvex: false,
                 closed: true,
                 coordinates: [[0.7819,0.9103],[0.3887,0.4325],[0.3212,0.342],[0.2878,0.2936],[0.2677,0.2619],[0.2544,0.2388],[0.2451,0.2209],[0.2384,0.2065],[0.2335,0.1944],[0.2298,0.184],[0.2272,0.175],[0.2254,0.1669],[0.2241,0.1597],[0.2234,0.1532],[0.2231,0.1472],[0.2232,0.1417],[0.2237,0.1365],[0.2244,0.1317],[0.2255,0.1272],[0.2268,0.1229],[0.2283,0.1188],[0.2301,0.1149],[0.2321,0.1112],[0.2344,0.1076],[0.237,0.1041],[0.2397,0.1007],[0.2428,0.0975],[0.2462,0.0943],[0.2498,0.0912],[0.2538,0.0881],[0.2581,0.0851],[0.2629,0.0822],[0.2681,0.0792],[0.2738,0.0764],[0.2801,0.0735],[0.287,0.0706],[0.2947,0.0678],[0.3034,0.0649],[0.3131,0.062],[0.3241,0.0592],[0.3368,0.0562],[0.3516,0.0533],[0.3691,0.0503],[0.3903,0.0472],[0.4167,0.0441],[0.4507,0.0409],[0.497,0.0377],[0.566,0.0344],[0.6872,0.0316],[1.1994,0.0373],[1.9169,0.0881],[2.1827,0.1328],[2.3621,0.1791],[2.4966,0.2267],[2.603,0.2756],[2.6899,0.3255],[2.7621,0.3763],[2.823,0.4278],[2.8746,0.4799],[2.9184,0.5326],[2.9557,0.5857],[2.9872,0.6391],[3.0135,0.6929],[3.0352,0.7468],[3.0527,0.8009],[3.0662,0.8551],[3.0761,0.9093],[3.0825,0.9634],[3.0856,1.0175],[3.0855,1.0713],[3.0824,1.125],[3.0762,1.1784],[3.0672,1.2314],[3.0554,1.2839],[3.0406,1.3359],[3.0231,1.3874],[3.0027,1.4381],[2.9795,1.488],[2.9534,1.537],[2.9244,1.5849],[2.8924,1.6317],[2.8573,1.6771],[2.819,1.7209],[2.7774,1.7629],[2.7323,1.8028],[2.6835,1.8404],[2.6307,1.8752],[2.5736,1.9066],[2.5118,1.9342],[2.4446,1.9571],[2.3714,1.9742],[2.291,1.9841],[2.202,1.9847],[2.102,1.9729],[1.9874,1.9438],[1.8516,1.8886],[1.6809,1.7896],[1.4382,1.6002],[0.7819,0.9103]]
+            },
+            {
+                color: '#ff0000',
+                mirrorRegionInx: 1
+            },
+            {
+                color: '#00b5b5',
+                mirrorRegionInx: 0
             }
         ]
     }
